+++ app/assets/javascripts/application.js
@@ -11,6 +11,8 @@
 // about supported directives.
 //
 //= require contour
+//= require compatibility/array_prototype_index_of
+//= require compatibility/array_prototype_map
 //= require turbolinks
 //= require external/prettify
 
@@ -22,6 +24,7 @@
 //= require agreements
 //= require datasets
 //= require comments
+//= require tags
 //= require external/jquery.lazyload.js
 //= require external/lang-matlab
 //= require external/lang-r
@@ -31,3 +34,4 @@
 //= require variables
 
 //= require external/jquery.textcomplete
+//= require external/colorpicker