+++ app/models/topic_tag.rb
@@ -1,0 +1,6 @@
+class TopicTag < ActiveRecord::Base
+
+  belongs_to :topic
+  belongs_to :tag
+
+end